Dynamed.xml
DynaMed is a clinical information resource used by physicians to answer clinical questions quickly and easily at the point of care. It includes thousands of topics that provide clinicians and educators with the evidence-based information they need to make informed decisions. Topics are updated daily as new evidence becomes available.